pyproject.toml
[project]
name = "todorama"
version = "0.1.0"
description = "Todorama - TODO Service for AI Agents with project support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.0.0",
    "boto3>=1.28.0",
    "click>=8.1.0",
    "fastapi>=0.104.0",
    "httpx>=0.25.0",
    "moto>=4.2.0",
    "nats-py>=2.7.0",
    "numpy>=1.24.0",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.38.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "opentelemetry-instrumentation-httpx>=0.42b0",
    "opentelemetry-instrumentation-sqlite3>=0.42b0",
    "opentelemetry-sdk>=1.38.0",
    "prometheus-client>=0.19.0",
    "pydantic-settings>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "python-multipart>=0.0.20",
    "python-telegram-bot>=21.0.0",
    "redis>=5.0.0",
    "reportlab>=4.0.0",
    "slack-sdk>=3.27.0",
    "speechrecognition>=3.10.0",
    "strawberry-graphql>=0.215.0",
    "uvicorn[standard]>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.0.0",
    "black>=25.11.0",
    "ruff>=0.14.4",
    "mypy>=1.18.2",
]

# No build system needed for application - dependencies only

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
ignore_missing_imports = true
no_strict_optional = true
check_untyped_defs = true
no_implicit_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--cov=todorama",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=80",
    "-v",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "performance: Performance and benchmark tests",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["todorama"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/migrations/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
]
